feat: implement SubFlow Node, Mermaid visualization, and modularize pi agent example
