+++ examples/pi/Cargo.toml
@@ -12,3 +12,7 @@ clap = { version = "4.4", features = ["derive"] }
 uuid = { version = "1.8", features = ["v4", "fast-rng"] }
 strum = { version = "0.26", features = ["derive"] }
 async-trait = "0.1"
+serde = { version = "1.0", features = ["derive"] }
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
+toml = "0.8"
+tracing = "0.1"
