+++ examples/pi/src/utils/mod.rs
@@ -1,0 +1,8 @@
+pub mod pi_llm;
+pub mod session_manager;
+pub mod tools;
+pub mod config;
+
+pub use pi_llm::PiLLM;
+pub use session_manager::SessionManager;
+pub use config::AppConfig;
