+++ examples/pi/src/utils/session_manager.rs
