+++ src/utils/mod.rs
@@ -4,7 +4,3 @@ pub mod text_chunking;
 pub mod vector_db;
 pub mod viz_debug;
 pub mod web_search;
-pub mod pi_llm;
-pub mod session_manager;
-pub mod tools;
-pub mod config;
